implemented ThreadPoolTaskQueue, although it only sometimes works (it works whenever sequencialTaskQueue decides to work)
git-svn-id: https://svn.ece.auckland.ac.nz/svn/se751-2012-g2/trunk@20 faf831e5-a2bb-0410-845a-b99f761cfef8

+++ ImageFilter/ImageFilter/Image.cpp
@@ -15,7 +15,7 @@ Image::Image(CGImageRef cgImg) {
     
     init(w, h, NULL);
     
-    CGRect rect = {{0,0},{_width,_height}}; 
+    CGRect rect = {{0,0},{static_cast<CGFloat>(_width),static_cast<CGFloat>(_height)}}; 
     CGContextDrawImage(_context, rect, cgImg); 
 }
